# protocol/rest/openapi.yaml
openapi: 3.1.0
info:
  title: SOBH Messenger API
  version: 1.0.0
  description: |
    REST surface of the SOBH platform. Every response uses the envelope from
    §66, every path is versioned under `/api/v1` (§67), and every authenticated
    call carries a Bearer access token.

    Endpoints marked **planned** are specified here but not yet implemented;
    they are listed so client work can proceed against a stable contract. See
    `docs/architecture/roadmap.md` for the delivery order.
  license:
    name: Proprietary
servers:
  - url: https://api.sobh.app
    description: Production
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/feature-flags:
    get:
      tags: [Platform]
      summary: List feature flags
      description: |
        Anonymous callers see the global state. Signed-in callers see their own
        bucketed answer, so the client can hide a feature instead of letting the
        user hit an error.
      security: [{}, { bearerAuth: [] }]
      responses:
        "200":
          description: Current flags
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: object
                        properties:
                          flags:
                            type: array
                            items: { $ref: "#/components/schemas/FeatureFlag" }
...
  /api/v1/chats:
    get:
      tags: [Chats]
      summary: List the caller's chats
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
        - name: before
          in: query
          description: RFC3339 timestamp cursor from `meta.next_cursor`.
          schema: { type: string, format: date-time }
        - name: folder_id
          in: query
          description: >
            Narrow the list to one chat folder. A folder is a view of the same
            list, so it is a parameter here rather than a second endpoint that
            would have to be kept in step with this one. Absent means the main
            list, which shows everything.
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Chat list, pinned first then most recently active
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: object
                        properties:
                          chats:
                            type: array
                            items: { $ref: "#/components/schemas/Chat" }
    post:
      tags: [Groups and channels]
      summary: Create a group or channel
      description: |
        The creator is made owner and the named members are added in the same
        transaction, so a chat is never left with no owner or with members and
        no creator.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, title]
              properties:
                type: { type: string, enum: [group, channel] }
                title: { type: string, maxLength: 128 }
                description: { type: string, maxLength: 512 }
                username:
                  type: string
                  description: Required for a public chat; it is how the chat is found and linked to.
                is_public: { type: boolean, default: false }
                member_ids:
                  type: array
                  maxItems: 200
                  items: { type: string, format: uuid }
      responses:
        "201":
          description: The new chat
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/Chat" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: That username is taken
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /api/v1/sync:
    get:
      tags: [Sync]
      summary: Fetch missed events
      description: |
        The REST form of `sync.request`. Returns durable events after `cursor`
        plus the current head, so a client that has been offline can catch up
        before opening a socket.
      parameters:
        - name: cursor
          in: query
          schema: { type: integer, format: int64, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, default: 200, maximum: 500 }
      responses:
        "200":
          description: Events and the server head
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: object
                        properties:
                          events:
                            type: array
                            items: { $ref: "#/components/schemas/SyncEvent" }
                          server_seq: { type: integer, format: int64 }
  /api/v1/contacts:
    get:
      tags: [Contacts]
      summary: List the caller's contacts
      description: |
        Last seen is present only when the contact's privacy rule allows this
        viewer to see it (§55).
      responses:
        "200":
          description: The address book
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: object
                        properties:
                          contacts:
                            type: array
                            items: { $ref: "#/components/schemas/Contact" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Contacts]
      summary: Add one contact by user id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
                first_name: { type: string, maxLength: 100 }
                last_name: { type: string, maxLength: 100 }
      responses:
        "200": { description: Added }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/contacts/sync:
    post:
      tags: [Contacts]
      summary: Match hashed phone numbers against accounts
      description: |
        The server never receives a phone number. The client normalises each to
        E.164, computes HMAC-SHA256 under the published pepper and uploads only
        the hex digests. `replace: true` makes the batch a full replacement, so
        a number deleted on the device disappears here too; a partial batch
        only adds. A user who has blocked the caller is not discoverable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                replace: { type: boolean, default: false }
                entries:
                  type: array
                  maxItems: 2000
                  items:
                    type: object
                    required: [digest]
                    properties:
                      digest:
                        type: string
                        description: Hex-encoded HMAC-SHA256, 64 characters
                        pattern: "^[0-9a-fA-F]{64}$"
                      first_name: { type: string }
                      last_name: { type: string }
      responses:
        "200":
          description: The entries that resolved to accounts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: object
                        properties:
                          matched:
                            type: array
                            items:
                              type: object
                              properties:
                                digest: { type: string }
                                user_id: { type: string, format: uuid }
                                display_name: { type: string }
                          uploaded: { type: integer }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/v1/me/data-requests:
    get:
      tags: [Users]
      summary: The caller's export and deletion requests
      responses:
        "200":
          description: The requests, newest first, and how long a deletion waits
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: object
                        properties:
                          deletion_delay_seconds: { type: integer }
                          requests:
                            type: array
                            items: { $ref: "#/components/schemas/DataRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Users]
      summary: Ask for a copy of your data, or for the account to be deleted
      description: >
        An export starts at once — there is nothing to reconsider about a copy
        of your own data — and produces a JSON document in the export bucket,
        recorded as media so it downloads through the ordinary presigned path.
        A deletion waits seven days first: somebody who deletes in anger, or
        whose account is deleted by an intruder holding a session, has that
        long to take it back. The account stays usable during the window,
        because revoking sessions immediately would remove the only means of
        cancelling.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [export, delete] }
      responses:
        "201":
          description: The request
        "409": { description: A request of that kind is already in progress }
...
  /api/v1/secret/chats:
    post:
      tags: [Secret chats]
      summary: Open the encrypted chat with a user
      description: |
        Returns the existing encrypted chat with that person or creates one, so
        calling it twice is safe and two devices racing land in one
        conversation rather than two.

        The peer must have at least one device that has published key material.
        A chat with someone who has none is refused rather than created: nothing
        sent into it could be encrypted to anybody, and a conversation that
        silently rejects every message is worse than a clear error.

        This is the only way to obtain a chat of type `secret`, which
        `POST /api/v1/secret/messages` requires (§24).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
      responses:
        "200":
          description: The encrypted chat
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: object
                        properties:
                          chat_id: { type: string, format: uuid }
                          created:
                            type: boolean
                            description: |
                              False when the chat already existed, so the caller
                              can tell a new conversation from an old one
                              without a second request.
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /api/v1/secret/messages:
    post:
      tags: [Secret chats]
      summary: Send ciphertext to one device
      description: |
        The server stores opaque bytes and cannot read them. Membership of the
        chat is checked so the mailbox is not an open relay; the payload itself
        is never examined. A repeated `client_message_id` returns the original
        envelope rather than storing a second one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chat_id, recipient_device_id, ciphertext, client_message_id]
              properties:
                chat_id: { type: string, format: uuid }
                recipient_device_id: { type: string, format: uuid }
                ciphertext:
                  type: string
                  description: Base64, at most 128 KiB decoded
                message_type: { type: integer, default: 1 }
                client_message_id: { type: string, format: uuid }
      responses:
        "201":
          description: Stored and announced
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: object
                        properties:
                          envelope_id: { type: string, format: uuid }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { description: The encrypted payload is too large for a message }
...
  /api/v1/stickers/search:
    get:
      tags: [Stickers]
      summary: Find sets by title, slug or the emoji their stickers carry
      parameters:
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: Matching sets, official ones first. Always a list, never null.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data:
                        type: object
                        properties:
                          sets:
                            type: array
                            items: { $ref: "#/components/schemas/StickerSet" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    botToken:
      type: http
      scheme: bearer
      description: |
        A bot token, presented as `Authorization: Bearer <token>`. The server
        stores only its SHA-256 hash, so a token cannot be recovered from a
        database dump and a lost one is replaced rather than looked up.
...
  responses:
    Unauthorized:
      description: Missing, expired or revoked token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
...
    ValidationFailed:
      description: The request body failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Envelope" }
...
  schemas:
    Envelope:
      type: object
      description: The single response shape used by every endpoint (§66).
      required: [success]
      properties:
        success: { type: boolean }
        data: { description: Present when `success` is true }
        error: { $ref: "#/components/schemas/ErrorBody" }
        meta: { $ref: "#/components/schemas/Meta" }
...
    FeatureFlag:
      type: object
      properties:
        key: { type: string }
        enabled: { type: boolean }
        rollout_percent: { type: integer }
        description: { type: string }
...